+++ falcosidekick-k8s-operator/src/state.py
@@ -6,18 +6,18 @@
 import itertools
 import logging
 from abc import ABC, abstractmethod
+from typing import Optional
 
 import ops
-from pydantic import BaseModel, ValidationError
+from charms.loki_k8s.v1.loki_push_api import LokiPushApiConsumer
+from pydantic import BaseModel, HttpUrl, ValidationError
 
+from certificates import TlsCertificateRequirer
 from config import CharmConfig, InvalidCharmConfigError
-from relations import InvalidLokiRelationError, LokiRelationManager
 
 logger = logging.getLogger(__name__)
 
-
-class InvalidStateError(Exception):
-    """Exception raised when a charm configuration is invalid and unrecoverable."""
+TlsHealthcheckPort = 2810  # Falcosidekick TLS healthcheck port (hardcoded)
 
 
 class CharmState(BaseModel):
@@ -28,15 +28,29 @@ class CharmState(BaseModel):
 
     Attributes:
         falcosidekick_listenport: The port on which Falcosidekick listens.
-        loki_endpoint_url: The URL of the Loki endpoint, if available.
+        falcosidekick_loki_endpoint: The URL of the Loki push API endpoint.
+        falcosidekick_loki_hostport: The host and port of the Loki push API endpoint.
+        falcosidekick_tlsserver_key_file: The file path to the TLS server private key.
+        falcosidekick_tlsserver_cert_file: The file path to the TLS server certificate.
+        falcosidekick_tlsserver_notlsport: The port for non-TLS connections (for health check only).
+        ca_cert: The CA certificate content for TLS verification.
     """
 
     falcosidekick_listenport: int
     falcosidekick_loki_endpoint: str
     falcosidekick_loki_hostport: str
+    falcosidekick_tlsserver_key_file: str
+    falcosidekick_tlsserver_cert_file: str
+    falcosidekick_tlsserver_notlsport: int
+    ca_cert: str
 
     @classmethod
-    def from_charm(cls, charm: ops.CharmBase, loki_relation: LokiRelationManager) -> "CharmState":
+    def from_charm(
+        cls,
+        charm: ops.CharmBase,
+        loki_push_api_consumer: LokiPushApiConsumer,
+        tls_certificate_requirer: TlsCertificateRequirer,
+    ) -> "CharmState":
         """Create a CharmState from a charm instance.
 
         Loads and validates the charm configuration, then constructs a CharmState
@@ -44,31 +58,42 @@ class CharmState(BaseModel):
 
         Args:
             charm: The charm instance from which to extract state.
-            loki_relation: The LokiRelationManager instance to get Loki relation data.
+            loki_push_api_consumer: The LokiPushApiConsumer instance to get Loki relation data.
+            tls_certificate_requirer: The TlsCertificateRequirer instance to get TLS data.
 
         Returns:
             CharmState: A validated CharmState instance.
 
         Raises:
             InvalidCharmConfigError: If configuration validation fails.
+            InvalidStateError: If relation data is invalid.
         """
         try:
             charm_config = charm.load_config(CharmConfig)
-            _url = loki_relation.get_loki_http_url()
-            loki_endpoint = _url.path if _url else "/loki/api/v1/push"
+            _url = _get_loki_ingress_endpoint(loki_push_api_consumer)
+            loki_endpoint = _url.path if _url and _url.path else "/loki/api/v1/push"
             loki_hostport = f"{_url.scheme}://{_url.host}:{_url.port}" if _url else ""
+            certificate_ready = tls_certificate_requirer.is_ready()
+            private_key_file = ""
+            certificate_file = ""
+            notlsport = charm_config.port  # default to listenport
+            if certificate_ready:
+                private_key_file = tls_certificate_requirer.private_key_name
+                certificate_file = tls_certificate_requirer.certificate_name
+                notlsport = TlsHealthcheckPort  # Falcosidekick TLS healthcheck port (hardcoded)
         except ValidationError as e:
             logger.error("Configuration validation error: %s", e)
             error_fields = set(itertools.chain.from_iterable(err["loc"] for err in e.errors()))
             error_field_str = " ".join(f"{f}" for f in error_fields)
             raise InvalidCharmConfigError(f"Invalid charm configuration: {error_field_str}") from e
-        except InvalidLokiRelationError as e:
-            logger.error("Loki relation data validation error: %s", e)
-            raise InvalidStateError("Invalid Loki relation data") from e
         return cls(
             falcosidekick_listenport=charm_config.port,
             falcosidekick_loki_endpoint=loki_endpoint,
             falcosidekick_loki_hostport=loki_hostport,
+            falcosidekick_tlsserver_key_file=private_key_file,
+            falcosidekick_tlsserver_cert_file=certificate_file,
+            falcosidekick_tlsserver_notlsport=notlsport,
+            ca_cert=tls_certificate_requirer.get_ca_cert(),
         )
 
 
@@ -97,3 +122,21 @@ class CharmBaseWithState(ops.CharmBase, ABC):
         Args:
             _: The hook event that triggered reconciliation.
         """
+
+
+def _get_loki_ingress_endpoint(loki_push_api_consumer: LokiPushApiConsumer) -> Optional[HttpUrl]:
+    """Get the first encounter Loki ingress endpoint.
+
+    Args:
+        loki_push_api_consumer: The LokiPushApiConsumer instance to get Loki relation data
+
+    Returns:
+        The first seen one Loki ingress endpoint in the relation data or None if not found.
+    """
+    try:
+        for endpoint in loki_push_api_consumer.loki_endpoints:
+            return HttpUrl(endpoint.get("url"))
+    except ValidationError:
+        logger.warning("Loki ingress endpoint not ready")
+
+    return None