+++ falcosidekick-k8s-operator/terraform/outputs.tf
@@ -9,5 +9,12 @@ output "app_name" {
 output "requires" {
   value = {
     send_loki_logs = "send-loki-logs"
+    certificates   = "certificates"
+  }
+}
+
+output "provides" {
+  value = {
+    http_endpoint = "http-endpoint"
   }
 }